+++ src/main/java/org/folio/entlinks/domain/entity/AuthoritySourceFileSource.java
@@ -1,0 +1,9 @@
+package org.folio.entlinks.domain.entity;
+
+public enum AuthoritySourceFileSource {
+  FOLIO,
+
+  LOCAL,
+
+  CONSORTIUM;
+}