+++ src/test/java/org/folio/entlinks/controller/AuthoritySourceFilesControllerIT.java
@@ -29,6 +29,7 @@ import org.folio.entlinks.domain.dto.AuthoritySourceFilePostDto;
 import org.folio.entlinks.domain.dto.AuthoritySourceFilePostDtoHridManagement;
 import org.folio.entlinks.domain.entity.AuthoritySourceFile;
 import org.folio.entlinks.domain.entity.AuthoritySourceFileCode;
+import org.folio.entlinks.domain.entity.AuthoritySourceFileSource;
 import org.folio.entlinks.exception.AuthoritySourceFileNotFoundException;
 import org.folio.entlinks.exception.RequestBodyValidationException;
 import org.folio.spring.test.extension.DatabaseCleanup;
@@ -56,7 +57,11 @@ class AuthoritySourceFilesControllerIT extends IntegrationTestBase {
   private static final Integer[] SOURCE_FILE_CODE_IDS = new Integer[] {1, 2, 3};
   private static final String[] SOURCE_FILE_CODES = new String[] {"c", "co", "cod"};
   private static final String[] SOURCE_FILE_NAMES = new String[] {"name1", "name2", "name3"};
-  private static final String[] SOURCE_FILE_SOURCES = {"local", "local", "folio"};
+  private static final AuthoritySourceFileSource[] SOURCE_FILE_SOURCES = new AuthoritySourceFileSource[] {
+    AuthoritySourceFileSource.LOCAL,
+    AuthoritySourceFileSource.LOCAL,
+    AuthoritySourceFileSource.FOLIO
+  };
   private static final String[] SOURCE_FILE_TYPES = new String[] {"type1", "type2", "type3"};
   private static final String[] SOURCE_FILE_URLS = new String[] {"baseUrl1", "baseUrl2", "baseUrl3"};
 
@@ -404,7 +409,7 @@ class AuthoritySourceFilesControllerIT extends IntegrationTestBase {
     var entity = new AuthoritySourceFile();
     entity.setId(SOURCE_FILE_IDS[sourceFileIdNum]);
     entity.setName(SOURCE_FILE_NAMES[sourceFileIdNum]);
-    entity.setSource(SourceEnum.FOLIO.getValue());
+    entity.setSource(AuthoritySourceFileSource.FOLIO);
     entity.setType(SOURCE_FILE_TYPES[sourceFileIdNum]);
     entity.setBaseUrl(SOURCE_FILE_URLS[sourceFileIdNum] + "/");