+++ src/test/java/org/folio/entlinks/service/authority/AuthoritySourceFileServiceTest.java
@@ -1,6 +1,7 @@
 package org.folio.entlinks.service.authority;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.folio.entlinks.domain.entity.AuthoritySourceFileSource.LOCAL;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -19,6 +20,7 @@ import java.util.UUID;
 import org.folio.entlinks.controller.converter.AuthoritySourceFileMapper;
 import org.folio.entlinks.domain.entity.AuthoritySourceFile;
 import org.folio.entlinks.domain.entity.AuthoritySourceFileCode;
+import org.folio.entlinks.domain.entity.AuthoritySourceFileSource;
 import org.folio.entlinks.domain.repository.AuthoritySourceFileRepository;
 import org.folio.entlinks.exception.AuthoritySourceFileNotFoundException;
 import org.folio.entlinks.exception.RequestBodyValidationException;
@@ -115,7 +117,7 @@ class AuthoritySourceFileServiceTest {
     code.setCode("code");
     var entity = new AuthoritySourceFile();
     entity.setAuthoritySourceFileCodes(Set.of(code));
-    entity.setSource("local");
+    entity.setSource(LOCAL);
     var expected = new AuthoritySourceFile(entity);
     when(repository.save(any(AuthoritySourceFile.class))).thenAnswer(i -> i.getArguments()[0]);
 
@@ -130,14 +132,14 @@ class AuthoritySourceFileServiceTest {
     var code = new AuthoritySourceFileCode();
     var entity = new AuthoritySourceFile();
     entity.setAuthoritySourceFileCodes(Set.of(code));
-    entity.setSource("folio");
+    entity.setSource(AuthoritySourceFileSource.FOLIO);
 
     var thrown = assertThrows(RequestBodyValidationException.class, () -> service.create(entity));
 
     verifyNoInteractions(repository);
     assertThat(thrown.getInvalidParameters()).hasSize(1);
     assertThat(thrown.getInvalidParameters().get(0).getKey()).isEqualTo("source");
-    assertThat(thrown.getInvalidParameters().get(0).getValue()).isEqualTo(entity.getSource());
+    assertThat(thrown.getInvalidParameters().get(0).getValue()).isEqualTo(entity.getSource().name());
   }
 
   @ParameterizedTest
@@ -147,7 +149,7 @@ class AuthoritySourceFileServiceTest {
     sourceFileCode.setCode(code);
     var entity = new AuthoritySourceFile();
     entity.setAuthoritySourceFileCodes(Set.of(sourceFileCode));
-    entity.setSource("local");
+    entity.setSource(LOCAL);
 
     var thrown = assertThrows(RequestBodyValidationException.class, () -> service.create(entity));
 
@@ -165,7 +167,7 @@ class AuthoritySourceFileServiceTest {
     code2.setCode("code2");
     var entity = new AuthoritySourceFile();
     entity.setAuthoritySourceFileCodes(Set.of(code1, code2));
-    entity.setSource("local");
+    entity.setSource(LOCAL);
 
     var thrown = assertThrows(RequestBodyValidationException.class, () -> service.create(entity));
 
@@ -181,7 +183,7 @@ class AuthoritySourceFileServiceTest {
     UUID id = UUID.randomUUID();
     entity.setId(id);
     entity.setName("updated name");
-    entity.setSource("updated source");
+    entity.setSource(LOCAL);
     var codeNew = new AuthoritySourceFileCode();
     codeNew.setCode("codeNew");
     entity.addCode(codeNew);
@@ -252,7 +254,7 @@ class AuthoritySourceFileServiceTest {
     UUID id = UUID.randomUUID();
     var authoritySourceFile = new AuthoritySourceFile();
     authoritySourceFile.setId(id);
-    authoritySourceFile.setSource("folio");
+    authoritySourceFile.setSource(AuthoritySourceFileSource.FOLIO);
 
     when(repository.findById(id)).thenReturn(Optional.of(authoritySourceFile));